Save given attachments in `addTechlogEntryAction` cloud function
- Attachments for a techlog entry are stored in Firebase storage here:
  `organizations/${organizationId}/aircrafts/${aircraftId}/techlog/${techlogEntryId}/${actionId}/${name}`
- The name of the file is a randomly generated UUID to prevent file name
  collisions (file extension is preserved). The original name is saved
  as custom metadata in the property `originalName`
